requirements.txt
google-generativeai==0.5.4
black==24.4.2
climage==0.2.0
python-dotenv~=1.0.1
pillow==10.3.0
